fix(cmd/gc): route wait dependency reads at the class binding (ga-ntixo)
Both wait dependency readers resolved only across work scopes — the
controller tick's leg set and the one-shot CLI's scope-candidate walk.
A binding-resident dependency missed on every leg, and a clean miss is
consumed as proof the dependency was deleted: a silent FailWait, no
event, and a session that sleeps forever.

Add the graph binding as a leg (typed, required) and probe it in the
CLI reader before the work scopes. Both arms are mutation-verified.

Also trims the split-store comment prose to two lines per site and
folds the class-routing tests into their existing sibling files.
