+++ cmd/gc/class_store_test.go
@@ -1,9 +1,11 @@
 package main
 
 import (
+	"context"
 	"testing"
 
 	"github.com/gastownhall/gascity/internal/beads"
+	"github.com/gastownhall/gascity/internal/molecule"
 )
 
 // controllerClassAccessor names a controllerState per-class accessor for the
@@ -124,4 +126,65 @@ func sameStorePtr(a, b beads.Store) bool {
 	ka, oka := storePointerKey(a)
 	kb, okb := storePointerKey(b)
 	return oka && okb && ka == kb
+}
+
+// TestCookOnClassRoutedPoursByRecipeClass pins the decision molecule.Cook cannot
+// make: it picks its store before the formula has compiled, so a one-store
+// caller pours graph-class workflows into the work ledger — and a poured v1
+// formula pushed the other way hides its steps from `gc hook`.
+func TestCookOnClassRoutedPoursByRecipeClass(t *testing.T) {
+	dir := convergenceResidenceFormulaDir(t)
+
+	for _, tt := range []struct {
+		name    string
+		formula string
+		// wantGraph is whether the molecule must land in the graph store.
+		wantGraph bool
+	}{
+		{"vapor formula is graph class", convergenceVaporFormula, true},
+		{"poured v1 formula is work class", convergencePouredFormula, false},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			// Disjoint ID spaces: two fresh MemStores both mint gc-1, so
+			// counting the store that should be EMPTY is the only sound oracle.
+			work := beads.NewMemStore()
+			graph := beads.NewMemStoreFrom(1000, nil, nil)
+
+			// The parent lives in NEITHER store: Instantiate stamps ParentID and
+			// never reads it back, which is the production cross-store shape.
+			res, err := cookOnClassRouted(context.Background(), work, graph, tt.formula, []string{dir}, molecule.Options{
+				ParentID: "gc-parent-lives-in-a-third-store",
+			})
+			if err != nil {
+				t.Fatalf("cookOnClassRouted: %v", err)
+			}
+			if res == nil || res.Created == 0 {
+				t.Fatalf("cook created no beads; the residence assertions below would be vacuous")
+			}
+
+			gotWork, gotGraph := countBeads(t, work), countBeads(t, graph)
+			wantStore, otherStore := "graph", "work"
+			wantCount, otherCount := gotGraph, gotWork
+			if !tt.wantGraph {
+				wantStore, otherStore = otherStore, wantStore
+				wantCount, otherCount = otherCount, wantCount
+			}
+			if wantCount != res.Created {
+				t.Errorf("the %s store holds %d of the %d beads the cook created; the molecule did not land in its class store", wantStore, wantCount, res.Created)
+			}
+			if otherCount != 0 {
+				t.Errorf("the %s store holds %d beads; the cook leaked across the class boundary", otherStore, otherCount)
+			}
+		})
+	}
+}
+
+// TestCookOnClassRoutedRequiresAParent keeps molecule.CookOn's attach-only
+// contract: a missing ParentID would detach a molecule no autoclose can reap.
+func TestCookOnClassRoutedRequiresAParent(t *testing.T) {
+	dir := convergenceResidenceFormulaDir(t)
+	_, err := cookOnClassRouted(context.Background(), beads.NewMemStore(), beads.NewMemStore(), convergencePouredFormula, []string{dir}, molecule.Options{})
+	if err == nil {
+		t.Fatal("cooking with no ParentID succeeded; the attach-only contract is gone")
+	}
 }
