+++ cmd/gc/molecule_autoclose.go
@@ -76,9 +76,8 @@ func doMoleculeAutoclose(beadID string, stdout, stderr io.Writer) {
 	// the city and every rig, and derive the store-ref from that store, so
 	// rig-store closes autoclose their molecule roots instead of silently
 	// no-op'ing (#3411).
-	// The molecule/workflow root this resolves is ClassGraph wherever the
-	// just-closed bead lives; see doWispAutoclose for why the graph leg routes
-	// off the owning store rather than following it.
+	// The root is ClassGraph wherever the just-closed bead lives, so the graph
+	// leg routes off the owning store rather than following it.
 	routeCfg, _ := loadCityConfigWithoutBuiltinPackRefresh(cityPath, io.Discard)
 	if store, dir, ok := autocloseOwningStore(beadID, cityPath); ok {
 		doMoleculeAutocloseWith(store, autocloseStoreRef(dir, cityPath), rec, beadID, stdout, cliGraphStore(store, routeCfg, cityPath))
@@ -120,12 +119,8 @@ func autocloseStoreRef(storeRoot, cityPath string) string {
 // owns it) and resolves/closes its molecule or graph-workflow root through the
 // graph-class store. A closed bead can be a work/source bead in a rig store
 // while the molecule root it belongs to lives in the graph store, so the
-// source-bead reverse scan and the root walk run on the graph store.
-//
-// The graph store is a REQUIRED, class-typed parameter for the reason given on
-// doWispAutocloseWith: as a variadic with a collapse-to-store default, both
-// hook-path callers omitted it and the only caller that runs in production read
-// the wrong store.
+// source-bead reverse scan and the root walk run on the graph store. It is a
+// required parameter for the reason given on doWispAutocloseWith.
 func doMoleculeAutocloseWith(store beads.Store, storeRef string, rec events.Recorder, beadID string, stdout io.Writer, graphClassStore beads.GraphStore) {
 	// Unwrapped for internal use; see doWispAutocloseWith.
 	graphStore := graphClassStore.Store
